refactor tests for building with go1.22
